refactor: extract locator directive parsing
